fix(previousprime): improve solution and subject
