+++ tests/previousprime_test/Cargo.toml
@@ -2,7 +2,7 @@
 name = "previousprime_test"
 version = "0.1.0"
 authors = ["OGordoo <marcosafonso8@hotmail.com>"]
-edition = "2021"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
